improve zshrc and zprofile

+++ install.sh
@@ -22,16 +22,19 @@ deploy() {
     symlink "vimrc"
     symlink "vim"
 
-    export XDG_CONFIG_HOME=$XDG_CONFIG_HOME
     mkdir -p $XDG_CONFIG_HOME
-    mkdir -p $XDG_CONFIG_HOME/fish
     mkdir -p $XDG_CONFIG_HOME/fish/functions
+    mkdir -p $XDG_CONFIG_HOME/zsh/functions
     ln -sf "$dotfiles/nvim" $XDG_CONFIG_HOME
     ln -sf $dotfiles/fish/config.fish $XDG_CONFIG_HOME/fish/config.fish
     ln -sf $dotfiles/fish/fishfile $XDG_CONFIG_HOME/fish/fishfile
     {
         cd $XDG_CONFIG_HOME/fish/functions
         ls -1 $dotfiles/fish/functions | xargs -I{} ln -sf $dotfiles/fish/functions/{} $XDG_CONFIG_HOME/fish/functions
+    }
+    {
+        cd $XDG_CONFIG_HOME/zsh/functions
+        ls -1 $dotfiles/zsh/functions | xargs -I{} ln -sf $dotfiles/zsh/functions/{} $XDG_CONFIG_HOME/zsh/functions
     }
 
     symlink "tmux.conf"
@@ -47,21 +50,29 @@ init() {
         git config --global alias.d diff
         git config --global alias.unstage "reset HEAD"
     fi
+
     if has zsh && [ ! -d "$HOME/.zplug" ]; then
         curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh| zsh
     fi
+
     if has fish && [ ! -f "$XDG_CONFIG_HOME/fish/functions/fisher.fish" ]; then
         curl -Lo $XDG_CONFIG_HOME/fish/functions/fisher.fish --create-dirs https://git.io/fisher
     fi
+
     if has git && has tmux && [ ! -d "$HOME/.tmux/plugins/tpm" ]; then
         git clone https://github.com/tmux-plugins/tpm $HOME/.tmux/plugins/tpm
     fi
-    {
-        echo 'set -x NODEBREW_ROOT $HOME/.nodebrew'
-        echo 'set NVIM /usr/share/nvim'
-        echo 'set -x NVIM $NVIM'
-    } > $XDG_CONFIG_HOME/fish/config.local.fish
-    echo 'export NVIM=/usr/share/nvim' > $HOME/.zshrc.local
+
+    if [ ! -f  $XDG_CONFIG_HOME/fish/config.local.fish ]; then
+        {
+            echo 'set -x NODEBREW_ROOT $HOME/.nodebrew'
+            echo 'set NVIM /usr/share/nvim'
+            echo 'set -x NVIM $NVIM'
+        } > $XDG_CONFIG_HOME/fish/config.local.fish
+    fi
+    if [ ! -f  $HOME/.zshrc.local ]; then
+        echo 'export NVIM=/usr/share/nvim' > $HOME/.zshrc.local
+    fi
 }
 if [ $# -eq 0 ]; then
     setup
